Add API app

+++ lib/bookshelf/interactors/book/destroy.rb
@@ -7,18 +7,29 @@ class Book::Destroy
 
   def initialize(id)
     @book = BookRepository.find(id)
-    @activity = Activity.new(
+  end
+
+  def call
+    if @book
+      @activity = build_activity
+
+      BookRepository.transaction do
+        BookRepository.delete(@book)
+        @activity = ActivityRepository.create(@activity)
+      end
+    else
+      fail!
+    end
+  end
+
+  private
+
+  def build_activity
+    Activity.new(
       timestamp: Time.now,
       action: Activity::Action::DELETE,
       old_title: @book.title,
       old_author: @book.author
     )
-  end
-
-  def call
-    BookRepository.transaction do
-      BookRepository.delete(@book)
-      @activity = ActivityRepository.create(@activity)
-    end
   end
 end
